Replace reader with empty reader
リーダ書き直しのため旧リーダを対比してコンパイラとVMをコメントアウト。
文字列に手を出してしまった時点でもう逃げられない
