+++ src/main.cpp
@@ -32,24 +32,28 @@ int main()
   std::list<std::string> history {""};
   std::size_t index {0};
 
-  for (std::string buffer {}; std::cout << "> ", std::getline(std::cin, buffer); ) try
+  while (std::cin) try
   {
-    std::string code {std::begin(buffer), boost::find(buffer, ';')};
-    std::cerr << "\x1B[38;5;240m" << ++index << " " << buffer << "\x1B[0m" << (std::empty(buffer) ? "\r" : "\n");
-
-    if (auto tokens {meevax::system::tokenize<std::list>(history.back() += code)}; not std::empty(tokens) and boost::count(tokens, "(") <= boost::count(tokens, ")"))
-    {
-      auto expression {read(tokens)};
-      // std::cerr << "[debug] read: " << expression << std::endl;
-      std::cerr << expression << std::endl;
-
-      auto machine_code {compile(expression)};
-      // std::cerr << "[debug] compile: " << machine_code << std::endl;
-
-      std::cout << machine.execute(machine_code) << "\n\n";
-      history.emplace_back("");
-    }
+    read(std::cin);
   }
+  // for (std::string buffer {}; std::cout << "> ", std::getline(std::cin, buffer); ) try
+  // {
+  //   std::string code {std::begin(buffer), boost::find(buffer, ';')};
+  //   std::cerr << "\x1B[38;5;240m" << ++index << " " << buffer << "\x1B[0m" << (std::empty(buffer) ? "\r" : "\n");
+  //
+  //   if (auto tokens {meevax::system::tokenize<std::list>(history.back() += code)}; not std::empty(tokens) and boost::count(tokens, "(") <= boost::count(tokens, ")"))
+  //   {
+  //     auto expression {read(tokens)};
+  //     // std::cerr << "[debug] read: " << expression << std::endl;
+  //     std::cerr << expression << std::endl;
+  //
+  //     auto machine_code {compile(expression)};
+  //     // std::cerr << "[debug] compile: " << machine_code << std::endl;
+  //
+  //     std::cout << machine.execute(machine_code) << "\n\n";
+  //     history.emplace_back("");
+  //   }
+  // }
   catch (const std::runtime_error& error)
   {
     std::cerr << "\x1B[31m[error] " << error.what() << "\x1B[0m\n\n";
